Add unstable flag --no-finished-line
Add a new unstable flag to hide the `Finished` line when cargo has
finished building, if everything was fresh.
The final `Finished` line can be a significant portion of the output
when invoking cargo multiple times via an external tool
(e.g. CMake + Ninja).
With this unstable flag the ninja progress bar is not interrupted if
cargo does not rebuild anything and the visible output is much nicer.